requirements.txt
fastapi
uvicorn
pydantic
Pillow
sqlalchemy
asyncpg
aiohttp
google-search-results
python-dotenv
pytest
httpx
pytest-asyncio
pytest-mock